Begin "starring" system (#10)
Allows users to star a series and have a more limited media list that will display media regardless of their status.

Currently the ability to switch between "Default" and "Starred" lists exists, however there is no way to star a media, nor does it account for media that don't meet the requirements for the default list.